# examples/basic/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Todo API",
    "version": "1.0.0",
    "description": "Generated by @developersailor/gea-openapi-decorators"
  },
  "paths": {
    "/todos": {
      "get": {
        "summary": "List all todos",
        "tags": [
          "Todos"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new todo",
        "tags": [
          "Todos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/todos/{id}": {
      "get": {
        "summary": "Get a todo by ID",
        "tags": [
          "Todos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a todo by ID",
        "tags": [
          "Todos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateTodoDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the todo item",
            "example": "Buy milk"
          },
          "description": {
            "type": "string",
            "description": "Detailed description",
            "example": "Organic whole milk"
          }
        },
        "required": [
          "title"
        ]
      },
      "TodoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique ID",
            "example": "1"
          },
          "title": {
            "type": "string",
            "description": "Title of the todo item",
            "example": "Buy milk"
          },
          "completed": {
            "type": "boolean",
            "description": "Whether the item is completed"
          }
        },
        "required": [
          "id",
          "title",
          "completed"
        ]
      }
    }
  }
}