# examples/basic/src/routes.generated.ts
// Auto-generated by @developersailor/gea-openapi-decorators. DO NOT EDIT.
// Zero-reflection, Ahead-of-Time static Hono routes for GeaStack C++ compilation.

import type { Hono } from 'hono';
import { TodoController } from './controllers/todo.controller.js';

export function registerRoutes(app: Hono): void {
  const todoController = new TodoController();

  app.get('/todos', async (c) => {
    const query_search = c.req.query('search');
    const result = todoController.list(query_search);
    if (result instanceof Response) {
      return result;
    }
    if (result === undefined || result === null) {
      return c.body(null, 204);
    }
    return c.json(result, 200);
  });
  app.get('/todos/:id', async (c) => {
    const param_id = c.req.param('id');
    const result = todoController.getById(param_id);
    if (result instanceof Response) {
      return result;
    }
    if (result === undefined || result === null) {
      return c.body(null, 204);
    }
    return c.json(result, 200);
  });
  app.post('/todos', async (c) => {
    const body_body = await c.req.json();
    const result = todoController.create(body_body);
    if (result instanceof Response) {
      return result;
    }
    if (result === undefined || result === null) {
      return c.body(null, 201);
    }
    return c.json(result, 201);
  });
  app.delete('/todos/:id', async (c) => {
    const param_id = c.req.param('id');
    const result = todoController.delete(param_id);
    if (result instanceof Response) {
      return result;
    }
    if (result === undefined || result === null) {
      return c.body(null, 204);
    }
    return c.json(result, 204);
  });
}